Revert "Some warnings"
This reverts commit 468d8effc8245557aa49104d264d6ba048f91a7a.

+++ src/ifcgeom/taxonomy.h
@@ -100,7 +100,7 @@ struct eigen_base {
 		o << std::string(indent, ' ') << class_name;
 		if (this->components_) {
 			int n = T::RowsAtCompileTime * T::ColsAtCompileTime;
-			for (int i = 0; i < n; ++i) {
+			for (size_t i = 0; i < n; ++i) {
 				o << " " << (*components_)(i);
 			}
 		}
